+++ dd-trace-core/src/test/groovy/datadog/trace/core/propagation/B3HttpExtractorTest.groovy
@@ -28,8 +28,8 @@ class B3HttpExtractorTest extends DDSpecification {
     final ExtractedContext context = extractor.extract(headers, MapGetter.INSTANCE)
 
     then:
-    context.traceId == traceId
-    context.spanId == spanId
+    context.traceId == traceId.longValue()
+    context.spanId == spanId.longValue()
     context.baggage == [:]
     context.tags == ["some-tag": "my-interesting-info"]
     context.samplingPriority == expectedSamplingPriority
@@ -56,8 +56,8 @@ class B3HttpExtractorTest extends DDSpecification {
 
     then:
     if (expectedTraceId) {
-      assert context.traceId == expectedTraceId
-      assert context.spanId == expectedSpanId
+      assert context.traceId == expectedTraceId.longValue()
+      assert context.spanId == expectedSpanId.longValue()
     } else {
       assert context == null
     }
